migrating test of parameter validation to run on a primitive
