+++ recipe_system/reduction/coreReduce.py
@@ -134,71 +134,6 @@ class Reduce:
         self._output_filenames = None
         self.recipename = args.recipename if args.recipename else '_default'
 
-    def _validate_uparms(self):
-        """
-        Validate the user parameters.
-
-        This is an internal method that checks the user parameters for
-        any obvious issues.  This lets us fail fast in case the user
-        made an error.
-
-        :raises: :class:~recipe_system.reduction.coreReduce.UnrecognizedParameterException: \
-            when a user parameter uses an unrecognized primitive or parameter name
-        """
-        if self.uparms:
-            for uparm in self.uparms:
-                key = uparm[0]
-                primitive = None
-                if ':' in key:
-                    split_key = key.split(':')
-                    if len(split_key) != 2:
-                        raise UnrecognizedParameterException("Expecting parameter or primitive:parameter in "
-                                                             "-p user parameters")
-                    primitive = split_key[0]
-                    key = split_key[1]
-                found_primitive = False
-                found_key = False
-                alternative_keys = list()
-                alternative_primitives = list()
-
-                def big_generator(base_class=config.Config):
-                    for cfg_class in base_class.__subclasses__():
-                        yield cfg_class
-                        for cfg_subclass in big_generator(cfg_class):
-                            yield cfg_subclass
-
-                for cfg_class in big_generator():
-                    if (primitive is None or f"{primitive}Config" == cfg_class.__name__) and \
-                            not cfg_class.__name__.startswith('core_'):
-                        found_primitive = True
-                        if primitive is None or f"{primitive}Config" == cfg_class.__name__:
-                            # We have to use _fields to avoid having to instantiate the config here
-                            for field_name, field_typ in cfg_class._fields.items():  # pylint: disable=protected-access
-                                if key == field_name:
-                                    found_key = True
-                                elif key.upper() == field_name.upper() and field_name not in alternative_keys:
-                                    alternative_keys.append(field_name)
-                    if primitive is not None:
-                        if f"{primitive.upper()}CONFIG" == cfg_class.__name__.upper() \
-                                and cfg_class.__name__[0:-6] not in alternative_primitives:
-                            alternative_primitives.append(cfg_class.__name__[0:-6])
-                if primitive and not found_primitive:
-                    if alternative_primitives:
-                        if len(alternative_primitives) == 1:
-                            raise UnrecognizedParameterException(f"Primitive {primitive} not found, did you mean {alternative_primitives[0]}?")
-                        else:
-                            raise UnrecognizedParameterException(f"Primitive {primitive} not found, did you mean one of {alternative_primitives}?")
-                    else:
-                        raise UnrecognizedParameterException(f"Primitive {primitive} not found")
-                if not found_key:
-                    if alternative_keys:
-                        if len(alternative_keys) == 1:
-                            raise UnrecognizedParameterException(f"Parameter {key} not found, did you mean {alternative_keys[0]}?")
-                        else:
-                            raise UnrecognizedParameterException(f"Parameter {key} not found, did you mean one of these? {alternative_keys}?")
-                    else:
-                        raise UnrecognizedParameterException(f"Parameter {key} not recognized")
-
     @property
     def upload(self):
         return self._upload
@@ -300,7 +235,6 @@ class Reduce:
             log.stdinfo("Found '{}' as a primitive.".format(pname))
             self._logheader(pname)
             try:
-                self._validate_uparms()
                 primitive_as_recipe()
             except Exception as err:
                 log_traceback(log)
@@ -309,7 +243,6 @@ class Reduce:
         else:
             self._logheader(recipe)
             try:
-                self._validate_uparms()
                 recipe(p)
             except Exception:
                 log.error("Reduce received an unhandled exception. Aborting ...")
